// extensions/cart-transformer-ts/generated/api.ts
export type CartOperation =
  /** A cart line expand operation. */
  { expand: ExpandOperation; merge?: never; update?: never; }
  |  /** A cart line merge operation. */
  { expand?: never; merge: MergeOperation; update?: never; }
  |  /** A cart line update operation. Only stores on the Shopify Plus plan can use apps with update operations. */
  { expand?: never; merge?: never; update: UpdateOperation; };
// ...
export type ExpandOperation = {
  /** The cart line id to expand. */
  cartLineId: Scalars['ID']['input'];
  /** The cart items to expand. */
  expandedCartItems: Array<ExpandedItem>;
  /** The image of the group. */
  image?: InputMaybe<ImageInput>;
  /** The price adjustment to the group. */
  price?: InputMaybe<PriceAdjustment>;
  /** Title override. If title is not provided, variant title is used. */
  title?: InputMaybe<Scalars['String']['input']>;
};
// ...
export type FunctionRunResult = {
  /** Cart operations to run on Cart. */
  operations: Array<CartOperation>;
};
// ...
export type MergeOperation = {
  /** The CartLine attributes to be added to the parent line. */
  attributes?: InputMaybe<Array<AttributeOutput>>;
  /** The list of cart lines to merge. */
  cartLines: Array<CartLineInput>;
  /** The image of the group. */
  image?: InputMaybe<ImageInput>;
  /** The product variant that models the group of lines. */
  parentVariantId: Scalars['ID']['input'];
  /** The price adjustment to the group. */
  price?: InputMaybe<PriceAdjustment>;
  /** The name of the group of lines to merge. If it isn't specified, it will use the parent_variant' name. */
  title?: InputMaybe<Scalars['String']['input']>;
};
// ...
export type UpdateOperation = {
  /** The ID of the cart line. */
  cartLineId: Scalars['ID']['input'];
  /** The image override for the cart line item. */
  image?: InputMaybe<ImageInput>;
  /** The price adjustment for the cart line item. */
  price?: InputMaybe<UpdateOperationPriceAdjustment>;
  /** The title override for the cart line item. */
  title?: InputMaybe<Scalars['String']['input']>;
};